Menu rechulon

+++ CrunchThroughHell/Src/LoveEngine/Input.h
@@ -71,7 +71,7 @@ namespace LoveEngine {
 			bool isControllerButtonState(ControllerButtonState s);
 			bool controllerConected();
 			bool justClicked();
-			Utilities::Vector2<float> mousePosition();
+			Utilities::Vector2<int> mousePosition();
 			Utilities::Vector2<float> relativeMousePosition();
 			void mouseVisibility(bool inGame);
 		};
